+++ app/Http/Controllers/CustomerController.php
@@ -12,12 +12,18 @@ use Illuminate\Support\Facades\Redirect;
 
 class CustomerController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
         return Inertia::render(
             'Customers',
-            ['customers' => Customer::orderByDesc('id')->get()
-            ->map(function (Customer $customer) {
+            [
+            'customers' => Customer::when($request->has('search'), function ($query) use ($request) {
+                $query->where('first_name', 'LIKE', '%' . $request->search . '%')
+                ->orWhere('last_name', 'LIKE', '%' . $request->search . '%');
+            })
+            ->orderByDesc('id')
+            ->paginate(25)
+            ->through(function (Customer $customer) {
                 return [
                     'first_name' => $customer->first_name,
                     'last_name' => $customer->last_name,
@@ -88,6 +94,7 @@ class CustomerController extends Controller
             'city' => $customer->city,
             'phone' => $customer->phone,
             'birth' => $customer->birth,
+            'fiscal_code' => $customer->fiscal_code,
             'postal_code' => $customer->postal_code,
             'province'=> $customer->province,
             'resident' => $customer->resident