+++ resources/js/Pages/Bookings.vue
@@ -1,0 +1,5 @@
+<template>
+    <div class="example-map">
+        <p class="example-text">aaa</p>
+    </div>
+</template>