+++ resources/lang/it/validation.php
@@ -114,7 +114,7 @@ return [
     'starts_with' => 'The :attribute must start with one of the following: :values.',
     'string' => 'The :attribute must be a string.',
     'timezone' => 'The :attribute must be a valid zone.',
-    'unique' => 'Campo :attribute presente',
+    'unique' => 'Il valore è già presente nel sistema',
     'uploaded' => 'The :attribute failed to upload.',
     'url' => 'Formato non valido per il campo :attribute',
     'uuid' => 'The :attribute must be a valid UUID.',